# Improve performance of AbstractController::Base#action_methods