### Motivation / Background

I was benchmarking some specs in my app, and saw this code come up in the memory_profiler. It is by no means the biggest memory allocation, but it is straightforward to make a slight improvement.



### Detail

Primarily, this saves allocating one array by using concat instead of + to add public_instance_methods(false).

That ends up being 10% less memory for my benchmark (3 actions), and 6% faster.




### Additional information
```
Calculating -------------------------------------
            original     8.352k memsize (   208.000  retained)
                        22.000  objects (     2.000  retained)
                         6.000  strings (     0.000  retained)
         refactored3     7.616k memsize (   408.000  retained)
                        11.000  objects (     7.000  retained)
                         3.000  strings (     3.000  retained)

Comparison:
         refactored3:       7616 allocated
            original:       8352 allocated - 1.10x more
Warming up --------------------------------------
            original     2.326k i/100ms
         refactored3     2.441k i/100ms
Calculating -------------------------------------
            original     23.336k (± 0.7%) i/s -    118.626k in   5.083658s
         refactored3     24.692k (± 1.2%) i/s -    124.491k in   5.042345s

Comparison:
            original:    23336.0 i/s
         refactored3:    24692.5 i/s - 1.06x  faster
```


Benchmark and results are also posted to https://gist.github.com/technicalpickles/4a4ae6a9e2c42963af43a89f75e768fe

See https://github.com/rails/rails/pull/49016 for my previous attempt.

### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Changes that are unrelated should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [x] Tests are added or updated if you fix a bug or add a feature.
* [x] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.