@technicalpickles 2023-08-24
Thanks!

> [Fix comments and indentation](https://github.com/rails/rails/pull/49029/commits/dc3da021ee35150c44f4ccd2be579f8a89a2ccd5)

fwiw, I did run rubocop, and it didn't seem opinionated about the formatting 🤔